twitter/top5volume/requirements.txt
tweepy~=3.10.0
requests~=2.25.1
numpy~=1.20.1
Pillow~=8.1.1
boto3~=1.17.60